requirements.txt
streamlit
pandas
numpy
matplotlib
streamlit-audiorec
scikit-learn
Pillow
plotly
pydub
openai
whisper
openai-whisper
soundfile
google-cloud-speech